Add PagePanelController
